backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.2

# LLM Framework
pyautogen==0.2.12
openai==1.12.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
werkzeug==3.0.1

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
chardet==5.2.0

# Document Parsing
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.1

# Data Validation
jsonschema==4.20.0

# MCP (Model Context Protocol)
mcp[cli]>=0.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
